Tidied up adjmat module -- this module now only contains adjacency matrix stuff
