+++ src/function/LessThan.cpp
@@ -22,6 +22,7 @@
 #include "LessThan.h"
 #include "FunctionShortcut.h"
 #include "FunctionOfVector.h"
+#include "FunctionOfMatrix.h"
 #include "core/ActionRegister.h"
 
 #include <cmath>
@@ -47,10 +48,21 @@ Use a switching function to determine how many components of the vector are less
 */
 //+ENDPLUMEDOC
 
+//+PLUMEDOC COLVAR LESS_THAN_MATRIX
+/*
+Transform all the elements of a matrix using a switching function that is one when the input value is smaller than a threshold
+
+\par Examples
+
+*/
+//+ENDPLUMEDOC
+
 typedef FunctionShortcut<LessThan> LessThanShortcut;
 PLUMED_REGISTER_ACTION(LessThanShortcut,"LESS_THAN")
 typedef FunctionOfVector<LessThan> VectorLessThan;
 PLUMED_REGISTER_ACTION(VectorLessThan,"LESS_THAN_VECTOR")
+typedef FunctionOfMatrix<LessThan> MatrixLessThan;
+PLUMED_REGISTER_ACTION(MatrixLessThan,"LESS_THAN_MATRIX")
 
 void LessThan::registerKeywords(Keywords& keys) {
   keys.add("compulsory","NN","6","The n parameter of the switching function ");
