+++ src/function/MoreThan.cpp
@@ -22,6 +22,7 @@
 #include "MoreThan.h"
 #include "FunctionShortcut.h"
 #include "FunctionOfVector.h"
+#include "FunctionOfMatrix.h"
 #include "core/ActionRegister.h"
 
 #include <cmath>
@@ -47,10 +48,21 @@ Use a switching function to determine how many of elements in the input vector a
 */
 //+ENDPLUMEDOC
 
+//+PLUMEDOC COLVAR MORE_THAN_MATRIX
+/*
+Transform all the elements of a matrix using a switching function that is one when the input value is larger than a threshold
+
+\par Examples
+
+*/
+//+ENDPLUMEDOC
+
 typedef FunctionShortcut<MoreThan> MoreThanShortcut;
 PLUMED_REGISTER_ACTION(MoreThanShortcut,"MORE_THAN")
 typedef FunctionOfVector<MoreThan> VectorMoreThan;
 PLUMED_REGISTER_ACTION(VectorMoreThan,"MORE_THAN_VECTOR")
+typedef FunctionOfMatrix<MoreThan> MatrixMoreThan;
+PLUMED_REGISTER_ACTION(MatrixMoreThan,"MORE_THAN_MATRIX")
 
 void MoreThan::registerKeywords(Keywords& keys) {
   keys.add("compulsory","NN","6","The n parameter of the switching function ");
