+++ src/matrixtools/Makefile
@@ -1,4 +1,4 @@
-USE=core tools matrixtools multicolvar
+USE=core tools 
 
 # generic makefile
 include ../maketools/make.module
