+++ src/matrixtools/MatrixTimesVector.cpp
@@ -19,7 +19,7 @@
    You should have received a copy of the GNU Lesser General Public License
    along with plumed.  If not, see <http://www.gnu.org/licenses/>.
 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ */
-#include "ActionWithMatrix.h"
+#include "core/ActionWithMatrix.h"
 #include "core/ActionRegister.h"
 
 //+PLUMEDOC MCOLVAR MATRIX_VECTOR_PRODUCT
@@ -32,7 +32,7 @@ Calculate the product of the matrix and the vector
 //+ENDPLUMEDOC
 
 namespace PLMD {
-namespace adjmat {
+namespace matrixtools {
 
 class MatrixTimesVector : public ActionWithMatrix {
 private:
