+++ src/symfunc/Fccubic.cpp
@@ -21,7 +21,7 @@
 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ */
 #include "function/FunctionTemplateBase.h"
 #include "function/FunctionShortcut.h"
-#include "adjmat/FunctionOfMatrix.h"
+#include "function/FunctionOfMatrix.h"
 #include "core/ActionRegister.h"
 
 #include <string>
@@ -100,7 +100,7 @@ public:
 
 typedef function::FunctionShortcut<Fccubic> FccubicShortcut;
 PLUMED_REGISTER_ACTION(FccubicShortcut,"FCCUBIC_FUNC")
-typedef adjmat::FunctionOfMatrix<Fccubic> MatrixFccubic;
+typedef function::FunctionOfMatrix<Fccubic> MatrixFccubic;
 PLUMED_REGISTER_ACTION(MatrixFccubic,"FCCUBIC_FUNC_MATRIX")
 
 void Fccubic::registerKeywords( Keywords& keys ) {
