+++ src/valtools/VStack.cpp
@@ -19,7 +19,7 @@
    You should have received a copy of the GNU Lesser General Public License
    along with plumed.  If not, see <http://www.gnu.org/licenses/>.
 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++ */
-#include "ActionWithMatrix.h"
+#include "core/ActionWithMatrix.h"
 #include "core/ActionRegister.h"
 
 //+PLUMEDOC MCOLVAR VSTACK
@@ -32,7 +32,7 @@ Create a matrix by stacking vectors together
 //+ENDPLUMEDOC
 
 namespace PLMD {
-namespace adjmat {
+namespace valtools {
 
 class VStack : public ActionWithMatrix {
 private:
